Updated getCurrentStockData to Use New API
